Scripts04/Program/requirements.txt
snowflake-connector-python[pandas]
psycopg2-binary
tqdm

